+++ SwiftChallenge3/Sources/Moduls/MainScreen/Views/Cell/MainSongCell.swift
@@ -108,7 +108,7 @@ final class MainSongCell: UITableViewCell {
         layout.scrollDirection = .horizontal
         
         songCollection = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        songCollection.register(SecondCollectionViewCell.self, forCellWithReuseIdentifier: SecondCollectionViewCell.reuseIdentifier)
+        songCollection.register(CollectionViewCell.self, forCellWithReuseIdentifier: CollectionViewCell.reuseIdentifier)
     }
     private func setupDelegate() {
         songCollection.delegate = self
@@ -124,7 +124,7 @@ extension MainSongCell {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .horizontal
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.register(SecondCollectionViewCell.self, forCellWithReuseIdentifier: SecondCollectionViewCell.reuseIdentifier)
+        collectionView.register(CollectionViewCell.self, forCellWithReuseIdentifier: CollectionViewCell.reuseIdentifier)
         return collectionView
     }
 }
@@ -156,7 +156,7 @@ extension MainSongCell: UICollectionViewDataSource, UICollectionViewDelegate {
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: SecondCollectionViewCell.reuseIdentifier, for: indexPath) as! SecondCollectionViewCell
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CollectionViewCell.reuseIdentifier, for: indexPath) as! CollectionViewCell
         
         cell.track = playlist?.tracks[indexPath.row]
         return cell
